:sparkles:feat:Completed sign out feature

+++ server/routes/todos.js
@@ -5,17 +5,18 @@ const Joi = require("joi");
 
 const router = express.Router();
 
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
   try {
     const todos = await Todo.find().sort({ date: -1 });
-    res.send(todos);
+    const filteredTodos = todos.filter((todo) => todo.uid === req.user._id);
+    res.send(filteredTodos);
   } catch (error) {
     res.status(500).send(error.message);
     console.log(error.message);
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3).max(30),
@@ -45,7 +46,7 @@ router.post("/", async (req, res) => {
     console.log(error.message);
   }
 });
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3).max(30),
@@ -60,6 +61,9 @@ router.put("/:id", async (req, res) => {
     const todo = await Todo.findById(req.params.id);
 
     if (!todo) return res.status(404).send("Todo not found...");
+
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo update failed...");
 
     const { name, author, isComplete, date, uid } = req.body;
 
@@ -81,11 +85,15 @@ router.put("/:id", async (req, res) => {
     console.log(error.message);
   }
 });
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
   try {
     const todo = await Todo.findById(req.params.id);
 
     if (!todo) return res.status(404).send("Todo not found...");
+
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo check/uncheck failed...");
+
     const updatedTodo = await Todo.findByIdAndUpdate(
       req.params.id,
       {
@@ -100,11 +108,14 @@ router.patch("/:id", async (req, res) => {
     console.log(error.message);
   }
 });
-router.delete("/:id", async (req, res) => {
+router.delete("/:id",auth, async (req, res) => {
   try {
     const todo = await Todo.findById(req.params.id);
 
     if (!todo) return res.status(404).send("Todo not found...");
+
+    if (todo.uid !== req.user._id)
+      return res.status(401).send("Todo check/uncheck failed...");
 
     const deletedtodo = await Todo.findByIdAndDelete(req.params.id);
     res.send(deletedtodo);
